Pass context to consistentRead
consistentRead now wraps the file reader with a context and will return
after context done in between reads. Couldn't seem to find a simple way
to go about interrupting a Read call. This implementation can't
interrupt in the middle of a Read call, but can still return early if
there are any attempts left, or in between Read operations.